requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cx-Oracle==8.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.42.0
django-filter==21.1
django-heroku==0.3.1
django-rest-knox==4.1.0
django-stubs==1.9.0
django-stubs-ext==0.3.1
django-taggit==1.3.0
djangorestframework==3.12.4
djangorestframework-jwt==1.7.2
djangorestframework-simplejwt==2.1
djangorestframework-stubs==1.4.0
Faker==5.8.0
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==3.0.2
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
oauthlib==3.1.1
Pillow==8.4.0
psutil==5.8.0
psycopg2==2.9.1
pycparser==2.20
Pygments==2.8.1
PyJWT==1.7.0
pylint==2.6.0
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.3
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
somepackage==1.2.3
sqlparse==0.4.1
tea==0.1.6
text-unidecode==1.3
toml==0.10.2
types-pytz==2021.3.0
types-PyYAML==6.0.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==3.0
uritemplate==4.1.1
urllib3==1.26.4
version==0.1.0
waitress==2.0.0
whitenoise==5.3.0
wrapt==1.12.1
